infra/src/requirements.txt
aws-cdk.core==1.129.0
aws-cdk.aws-cognito==1.129.0
aws-cdk.aws-elasticsearch==1.129.0
aws-cdk.aws-opensearchservice==1.129.0
aws-cdk.aws_codepipeline_actions==1.129.0
